one less class in the world
