--- src/main/java/org/cancogenvirusseq/muse/config/websecurity/AuthProperties.java
@@ -1,63 +1,0 @@
-/*
- * Copyright (c) 2021 The Ontario Institute for Cancer Research. All rights reserved
- *
- * This program and the accompanying materials are made available under the terms of the GNU Affero General Public License v3.0.
- * You should have received a copy of the GNU Affero General Public License along with
- * this program. If not, see <http://www.gnu.org/licenses/>.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY
- * EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES
- * OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT
- * SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
- * INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED
- * TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS;
- * OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER
- * IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
- * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-package org.cancogenvirusseq.muse.config.websecurity;
-
-import com.google.common.collect.ImmutableList;
-import java.util.List;
-import lombok.Data;
-import lombok.Getter;
-import lombok.Setter;
-import lombok.Value;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.boot.context.properties.ConstructorBinding;
-import org.springframework.context.annotation.Configuration;
-
-import javax.validation.constraints.NotNull;
-import javax.validation.constraints.Pattern;
-
-@Data
-@Configuration
-@ConfigurationProperties(prefix = "auth")
-public class AuthProperties {
-  String jwtPublicKeyUrl;
-
-  String jwtPublicKeyStr;
-
-  ScopesConfig scopes = new ScopesConfig();
-
-  @Getter
-  @Setter
-  public static class ScopesConfig {
-    private String system;
-    private final ScopesConfig.StudyScopeConfig study = new ScopesConfig.StudyScopeConfig();
-
-    @Getter
-    @Setter
-    public static class StudyScopeConfig {
-
-      @NotNull
-      @Pattern(regexp = "^\\w+\\W$")
-      private String prefix;
-
-      @NotNull
-      @Pattern(regexp = "^\\W\\w+$")
-      private String suffix;
-    }
-  }
-}
